+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Fixed
+ - Fix Proton prefix detection when the prefix directory is located inside a `SteamApps` directory instead of `steamapps`
+
 ## [1.3] - 2019-11-06
 ### Added
  - Non-Steam applications are now detected.