+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
  - Fix Proton prefix detection when the prefix directory is located inside a `SteamApps` directory instead of `steamapps`
  - Use the most recently used Proton prefix when multiple prefix directories are found for a single game
+ - Fix Python 3.5 compatibility
 
 ## [1.3] - 2019-11-06
 ### Added