+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [1.3.1] - 2019-11-21
 ### Fixed
  - Fix Proton prefix detection when the prefix directory is located inside a `SteamApps` directory instead of `steamapps`
  - Use the most recently used Proton prefix when multiple prefix directories are found for a single game
@@ -85,7 +85,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
  - The last release of protontricks maintained by [@Sirmentio](https://github.com/Sirmentio).
 
-[Unreleased]: https://github.com/Matoking/protontricks/compare/1.3...HEAD
+[Unreleased]: https://github.com/Matoking/protontricks/compare/1.3.1...HEAD
+[1.3.1]: https://github.com/Matoking/protontricks/compare/1.3...1.3.1
 [1.3]: https://github.com/Matoking/protontricks/compare/1.2.5...1.3
 [1.2.5]: https://github.com/Matoking/protontricks/compare/1.2.4...1.2.5
 [1.2.4]: https://github.com/Matoking/protontricks/compare/1.2.3...1.2.4